+++ chat/src/main/java/ru/job4j/chat/controller/RoleController.java
@@ -3,6 +3,7 @@ package ru.job4j.chat.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.model.Role;
 import ru.job4j.chat.repository.RoleRepository;
 
@@ -30,14 +31,20 @@ public class RoleController {
     @GetMapping("/{id}")
     public ResponseEntity<Role> findById(@PathVariable int id) {
         var role = this.roleRepository.findById(id);
+        if (role.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Role not found");
+        }
         return new ResponseEntity<>(
-                role.orElse(new Role()),
-                role.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
+                role.get(),
+                HttpStatus.OK
         );
     }
 
     @PostMapping("/")
     public ResponseEntity<Role> create(@RequestBody Role role) {
+        if (role.getName() == null) {
+            throw new NullPointerException("Name field is empty");
+        }
         Role buff = Role.of(role.getName());
         return new ResponseEntity<>(
                 this.roleRepository.save(buff),
@@ -47,6 +54,9 @@ public class RoleController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Role role) {
+        if (role.getName() == null) {
+            throw new NullPointerException("Name field is empty");
+        }
         Role buff = this.roleRepository.findById(role.getId()).get();
         buff.setName(role.getName());
         this.roleRepository.save(buff);
@@ -55,6 +65,9 @@ public class RoleController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (this.roleRepository.findById(id).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Role not found");
+        }
         Role role = new Role();
         role.setId(id);
         this.roleRepository.delete(role);
