+++ chat/src/main/java/ru/job4j/chat/controller/RoomController.java
@@ -1,11 +1,17 @@
 package ru.job4j.chat.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.model.Room;
 import ru.job4j.chat.repository.RoomRepository;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
 
 import java.util.stream.Collectors;
@@ -14,9 +20,13 @@ import java.util.stream.StreamSupport;
 @RestController
 @RequestMapping("/room")
 public class RoomController {
+
+    private final ObjectMapper mapper;
+
     private final RoomRepository roomRepository;
 
-    public RoomController(RoomRepository roomRepository) {
+    public RoomController(ObjectMapper mapper, RoomRepository roomRepository) {
+        this.mapper = mapper;
         this.roomRepository = roomRepository;
     }
 
@@ -30,14 +40,23 @@ public class RoomController {
     @GetMapping("/{id}")
     public ResponseEntity<Room> findById(@PathVariable int id) {
         var room = this.roomRepository.findById(id);
+        if (room.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Room not found");
+        }
         return new ResponseEntity<>(
-                room.orElse(new Room()),
-                room.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
+                room.get(),
+                HttpStatus.OK
         );
     }
 
     @PostMapping("/")
     public ResponseEntity<Room> create(@RequestBody Room room) {
+        if (room.getName() == null) {
+            throw new NullPointerException("Name field is empty");
+        }
+        if (room.getName().contains("stop-word")) {
+            throw new IllegalArgumentException("You cant use stop-word in room name");
+        }
         Room buff = Room.of(room.getName());
         return new ResponseEntity<>(
                 this.roomRepository.save(buff),
@@ -47,6 +66,9 @@ public class RoomController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Room room) {
+        if (room.getName() == null) {
+            throw new NullPointerException("Name field is empty");
+        }
         Room buff = this.roomRepository.findById(room.getId()).get();
         buff.setName(room.getName());
         this.roomRepository.save(buff);
@@ -55,9 +77,25 @@ public class RoomController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (this.roomRepository.findById(id).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Room not found");
+        }
         Room room = new Room();
         room.setId(id);
         this.roomRepository.delete(room);
         return ResponseEntity.ok().build();
+    }
+
+    @ExceptionHandler(value = IllegalArgumentException.class)
+    public void exceptionHandler(
+            Exception e,
+            HttpServletRequest req,
+            HttpServletResponse resp) throws IOException {
+        resp.setStatus(HttpStatus.BAD_REQUEST.value());
+        resp.setContentType("application/json");
+        resp.getWriter().write(mapper.writeValueAsString(new HashMap<>() { {
+            put("message", e.getMessage());
+            put("type", e.getClass());
+        }}));
     }
 }
