+++ chat/src/main/java/ru/job4j/chat/controller/UserController.java
@@ -5,6 +5,7 @@ import org.springframework.http.*;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.model.Person;
 import ru.job4j.chat.model.Role;
 import ru.job4j.chat.repository.PersonRepository;
@@ -42,9 +43,12 @@ public class UserController {
     @GetMapping("/{id}")
     public ResponseEntity<Person> findById(@PathVariable int id) {
         var person = this.personRepository.findById(id);
+        if (person.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Person not found");
+        }
         return new ResponseEntity<>(
-                person.orElse(new Person()),
-                person.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
+                person.get(),
+                HttpStatus.OK
         );
     }
 
@@ -62,6 +66,9 @@ public class UserController {
             @PathVariable("id") int id,
             @RequestBody Person person,
             @RequestHeader("Authorization") String token) {
+        if (person.getUsername() == null || person.getPassword() == null) {
+            throw new NullPointerException("Username or Password field's is empty");
+        }
         HttpHeaders headers = new HttpHeaders();
         headers.set("Authorization", token);
         HttpEntity<?> entity = new HttpEntity<>(headers);
@@ -78,6 +85,9 @@ public class UserController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Person person) {
+        if (person.getUsername() == null || person.getPassword() == null) {
+            throw new NullPointerException("Username or Password field's is empty");
+        }
         Person buff = this.personRepository.findById(person.getId()).get();
         buff.setUsername(person.getUsername());
         buff.setPassword(person.getPassword());
@@ -87,6 +97,9 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (this.personRepository.findById(id).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Person not found");
+        }
         Person person = new Person();
         person.setId(id);
         this.personRepository.delete(person);
